MCA-3 - Update java/spring versions and refactor code
