--- .eval/src/main/java/com/hackerrank/eval/evaluation/PerformEvaluation.java
@@ -1,13 +1,0 @@
-package com.hackerrank.eval.evaluation;
-
-public class PerformEvaluation {
-  @SuppressWarnings("java:S106")
-  public static void main(String[] args) {
-    HackerrankYaml hackerrankYaml = new HackerrankYaml();
-    TestResult testResult = new TestResult(hackerrankYaml);
-    TestEvaluator testEvaluator = new TestEvaluator(hackerrankYaml, testResult);
-
-    double score = testEvaluator.evaluate();
-    System.out.println(String.format("Candidate scored: %.2f", score));
-  }
-}
