--- .eval/src/main/java/com/hackerrank/eval/evaluation/Status.java
@@ -1,6 +1,0 @@
-package com.hackerrank.eval.evaluation;
-
-enum Status {
-  PASSED,
-  FAILED
-}
