--- .eval/src/main/resources/hackerrank.yml
@@ -1,41 +1,0 @@
-version: 1.0
-configuration:
-  has_webserver: true
-  readonly_paths:
-    - .eval/*
-  scoring:
-    command: .eval/score.sh
-    files:
-      - .eval/target/surefire-reports/TEST-com.hackerrank.eval.BestPracticeTests.xml
-      - .eval/target/surefire-reports/TEST-com.hackerrank.eval.FunctionalTests.xml
-      - .eval/target/surefire-reports/TEST-com.hackerrank.eval.TestCoverageTests.xml
-    test_total_score: 100 #Max points that can be obtained
-    testcase_weights:
-      # The weights must sum up to a total of 1
-      # Scoring guidance:
-      # 5 point = 0.05
-      # 10 points = 0.1
-      # 20 points = 0.2
-
-      # Best practices(20 marks)
-      statusCode404WhenNonExistentCovidRequested: 0.05
-      statusCode400WhenTop5ByInvalid: 0.1
-      statusCode400WhenTotalByInvalid: 0.05
-
-      # Functional tests(40 marks)
-      testTop5: 0.2
-      testTotal: 0.2
-
-      #dashboard test(40 marks)
-      correctDashboardReportReturned: 0.4
-
-      #placeholder test
-      wroteAtLeastXNumberOfTests{int}[1]: 0.0
-
-  ide_config:
-    default_open_files:
-      - src/main/java/com/hackerrank/api/Application.java
-    project_menu:
-      run: mvn clean spring-boot:run
-      install: mvn clean install
-      test: mvn clean test
