--- .eval/src/test/java/com/hackerrank/eval/model/Covid.java
@@ -1,26 +1,0 @@
-package com.hackerrank.eval.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import java.io.Serializable;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Covid implements Serializable {
-  private Long id;
-  private String country;
-  private String continent;
-  private Integer active;
-  private Integer death;
-  private Integer recovered;
-}
