+++ .gitignore
@@ -49,7 +49,6 @@ out/
 ## Eclipse
 ##############################
 .settings/
-bin/
 tmp/
 .metadata
 .classpath
@@ -66,7 +65,6 @@ local.properties
 ## NetBeans
 ##############################
 nbproject/private/
-build/
 nbbuild/
 dist/
 nbdist/
@@ -82,19 +80,4 @@ nb-configuration.xml
 ##############################
 ## OS X
 ##############################
-.DS_Store!/.github/
-!/.github/**
-!/.github/
-!/.github/**
-!/.github/
-!/.github/**
-!/.github/
-!/.github/**
-!/.github/
-!/.github/**
-!/.github/
-!/.github/**
-!/.github/
-!/.github/**
-!/.github/
-!/.github/**
+.DS_Store/
