--- PROJECT_FILES_INSTRUCTIONS.md
@@ -1,25 +1,0 @@
-## Read-only Files
-The following files are marked read-only. You cannot edit these files
-in the editor; however, it is possible from the terminal. You must not
-modify or delete these files because doing so results in a zero score.
-
-* .eval/.gitignore
-* .eval/integration-tests.sh
-* .eval/pom.xml
-* .eval/score.sh
-* .eval/src/main/java/com/hackerrank/eval/evaluation/HackerrankYaml.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/MavenSurefireSingleTestResult.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/MavenSurefireTestResult.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/PerformEvaluation.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/Status.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/TestEvaluator.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/TestResult.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/reports/TestCoverageCalculator.java
-* .eval/src/main/java/com/hackerrank/eval/evaluation/reports/TestCoverageReport.java
-* .eval/src/main/resources/hackerrank.yml
-* .eval/src/test/java/com/hackerrank/eval/BestPracticeTests.java
-* .eval/src/test/java/com/hackerrank/eval/FunctionalTests.java
-* .eval/src/test/java/com/hackerrank/eval/TestCoverageTests.java
-* .eval/src/test/java/com/hackerrank/eval/extensions/RESTExtension.java
-* .eval/src/test/java/com/hackerrank/eval/model/Covid.java
-* .eval/src/test/java/com/hackerrank/eval/model/Report.java
