+++ src/main/java/com/covid/api/exception/BadRequestException.java
@@ -1,0 +1,12 @@
+package com.covid.api.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class BadRequestException extends RuntimeException {
+
+  public BadRequestException(String message) {
+    super(message);
+  }
+}
