+++ src/main/java/com/covid/api/exception/ElementNotFoundException.java
@@ -1,0 +1,12 @@
+package com.covid.api.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ElementNotFoundException extends RuntimeException {
+
+  public ElementNotFoundException(String message) {
+    super(message);
+  }
+}
