+++ src/main/java/com/covid/api/model/Covid.java
@@ -1,0 +1,34 @@
+package com.covid.api.model;
+
+//import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import java.io.Serializable;
+
+//TODO: Add Dto for the view, and mappings to convert from/to.
+@Entity
+@Getter
+@Setter
+@Builder
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+//@JsonIgnoreProperties(ignoreUnknown = true)
+public class Covid implements Serializable {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+  private String country;
+  private String continent;
+  private Integer active;
+  private Integer death;
+  private Integer recovered;
+}
