--- src/main/java/com/hackerrank/api/controller/HealthController.java
@@ -1,18 +1,0 @@
-package com.hackerrank.api.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthController {
-  @GetMapping("/health")
-  public ResponseEntity<String> health() {
-    return ResponseEntity.ok().build();
-  }
-
-  @GetMapping("/ready")
-  public ResponseEntity<String> ready() {
-    return ResponseEntity.ok().build();
-  }
-}
