--- src/main/java/com/hackerrank/api/exception/ElementNotFoundException.java
@@ -1,12 +1,0 @@
-package com.hackerrank.api.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ElementNotFoundException extends RuntimeException {
-
-  public ElementNotFoundException(String message) {
-    super(message);
-  }
-}
