--- src/main/java/com/hackerrank/api/model/Covid.java
@@ -1,33 +1,0 @@
-package com.hackerrank.api.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import java.io.Serializable;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Entity
-public class Covid implements Serializable {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-  private String country;
-  private String continent;
-  private Integer active;
-  private Integer death;
-  private Integer recovered;
-}
