--- src/main/java/com/hackerrank/api/model/Report.java
@@ -1,22 +1,0 @@
-package com.hackerrank.api.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import java.io.Serializable;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Builder
-@ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Report implements Serializable {
-  private String continent;
-  private Double impactFactor;
-}
