--- src/test/java/com/hackerrank/api/controller/CovidControllerTest.java
@@ -1,22 +1,0 @@
-package com.hackerrank.api.controller;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@AutoConfigureMockMvc
-@Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD, scripts = "classpath:data.sql")
-class CovidControllerTest {
-
-  @Test
-  public void testCreation() throws Exception {
-    //placeholder test to trigger db script
-    Assertions.assertEquals(1, 1);
-  }
-}
